+++ SOSS/APPLESOSS/README.md
@@ -1,0 +1,6 @@
+## Code for the APPLESOSS&trade Empirical Profile Construction Module
+
+The empirical profile construction aka the *APPLESOSS* (A Producer of ProfiLEs for SOSS; Radica et al. in prep) module builds 2D spatial profiles for the first and second diffraction orders for a NIRISS/SOSS
+GR700XD/CLEAR observation. The profiles are entirely data driven, only employing models where absolutely necessary and as such retain a high level (hopefully...) of fidelity to the original observations.
+
+These profiles are intended to be used as the specprofile reference file for the *ATOCA* 1D extration algorithm.
