+++ SOSS/APPLESOSS/Ref_files/SOSS_PSFs/SOSS_os1_256x256_1.000000_0.fits
