+++ SOSS/APPLESOSS/Ref_files/width_coefficients.csv
